+++ template/src/bin/main.rs
@@ -42,6 +42,10 @@ use esp_backtrace as _;
 extern crate alloc;
 //ENDIF
 
+// This creates a default app-descriptor required by the esp-idf bootloader.
+// For more information see: <https://docs.espressif.com/projects/esp-idf/en/stable/esp32/api-reference/system/app_image_format.html#application-description>
+esp_bootloader_esp_idf::esp_app_desc!();
+
 #[main]
 fn main() -> ! {
     //REPLACE generate-version generate-version