+++ CHANGELOG.md
@@ -18,12 +18,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The tool now prints the selected options (#154)
 - Enable the `clippy::mem_forget` lint since mem::forget is generally not safe to use with esp-hal. (#161)
 - Added option to enable integration with the `trouble` BLE library (#179)
+- Added `esp-bootloader-esp-idf` package (#166)
 
 ### Changed
 
 - The visual style in certain terminals no longer uses emojis (#173)
 - Add a description to the version check output (#178)
 - `esp-hal` is now pinned to avoid updating to a new prerelease by accident (#186)
+- `esp-hal` updated and pinned to `beta.1` (#166)
 
 ### Fixed